requirements.txt
transformers>=4.30.0
torch>=2.0.0
datasets>=2.12.0
numpy>=1.24.0
tqdm>=4.65.0
pandas>=2.0.0
scikit-learn>=1.2.0
tensorboard>=2.13.0
python-dotenv>=1.0.0
huggingface-hub>=0.19.0 